+++ src/main/java/org/una/programmingIII/WikiPets/Service/ReviewServiceImplementation.java
@@ -1,15 +1,22 @@
 package org.una.programmingIII.WikiPets.Service;
 
+import org.jetbrains.annotations.NotNull;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.una.programmingIII.WikiPets.Dto.ReviewDto;
+import org.una.programmingIII.WikiPets.Exception.BlankInputException;
+import org.una.programmingIII.WikiPets.Exception.CustomException;
+import org.una.programmingIII.WikiPets.Exception.InvalidInputException;
+import org.una.programmingIII.WikiPets.Exception.NotFoundElementException;
 import org.una.programmingIII.WikiPets.Mapper.GenericMapper;
 import org.una.programmingIII.WikiPets.Mapper.GenericMapperFactory;
 import org.una.programmingIII.WikiPets.Model.*;
 import org.una.programmingIII.WikiPets.Repository.ReviewRepository;
 
-import java.util.List;
-import java.util.stream.Collectors;
+import java.time.LocalDate;
+import java.util.Map;
 
 @Service
 public class ReviewServiceImplementation implements ReviewService {
@@ -22,30 +29,68 @@ public class ReviewServiceImplementation implements ReviewService {
         this.reviewMapper = mapperFactory.createMapper(Review.class, ReviewDto.class);
     }
 
-    public List<ReviewDto> getAllReviews() {
-        List<Review> reviews = reviewRepository.findAll();
-        return reviews.stream().map(this.reviewMapper::convertToDTO).collect(Collectors.toList());
+    @Override
+    public Map<String, Object> getReviews(int page, int size) {
+        Page<Review> reviews = reviewRepository.findAll(PageRequest.of(page, size));
+        return Map.of("reviews", reviews.map(this::convertToDto).getContent(), "totalPages", reviews.getTotalPages(), "totalElements", reviews.getTotalElements());
     }
 
+
+    @Override
     public ReviewDto getReviewById(Long id) {
         Review review = reviewRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Review Not Found with id: " + id));
+                .orElseThrow(() -> new NotFoundElementException("Review Not Found with id: " + id));
         return reviewMapper.convertToDTO(review);
     }
 
-    public ReviewDto createReview(ReviewDto reviewDto) {
+    @Override
+    public ReviewDto createReview(@NotNull ReviewDto reviewDto) {
+        if (reviewDto.getComment().isBlank()) {
+            throw new BlankInputException("Comment cannot be empty");
+        }
+
+        if (reviewDto.getRating() > 10 || reviewDto.getRating() < 1) {
+            throw new BlankInputException("Review only accept 1-10 rating");
+        }
+        reviewDto.setLastUpdate(LocalDate.now());
+        reviewDto.setCreateDate(LocalDate.now());
         Review review = reviewMapper.convertToEntity(reviewDto);
-        Review savedReview = reviewRepository.save(review);
-        return reviewMapper.convertToDTO(savedReview);
+        return reviewMapper.convertToDTO(reviewRepository.save(review));
     }
 
-    public void deleteReview(Long id) {
+    @Override
+    public boolean deleteReview(Long id) {
+        if (id == null || id <= 0) {
+            return false;
+        }
+
+        if (!(reviewRepository.existsById(id))) {
+            throw new NotFoundElementException("Review not found with id: " + id);
+        }
         reviewRepository.deleteById(id);
+        return true;
+
     }
 
-    public ReviewDto updateReview(ReviewDto reviewDto) {
-        Review review = reviewMapper.convertToEntity(reviewDto);
-        Review updatedReview = reviewRepository.save(review);
-        return reviewMapper.convertToDTO(updatedReview);
+    @Override
+    public ReviewDto updateReview(@NotNull ReviewDto reviewDto) {
+        if (reviewDto.getComment().isBlank()) {
+            throw new BlankInputException("Comment cannot be empty");
+        }
+
+        if (reviewDto.getRating() > 10 || reviewDto.getRating() < 1) {
+            throw new BlankInputException("Review only accept 1-10 rating");
+        }
+        Review oldReview = reviewRepository.findById(reviewDto.getId())
+                .orElseThrow(() -> new NotFoundElementException("Review with id " + reviewDto.getId() + " not found"));
+
+        Review newReview = reviewMapper.convertToEntity(reviewDto);
+        newReview.setCreateDate(oldReview.getCreateDate());
+        newReview.setLastUpdate(LocalDate.now());
+        return reviewMapper.convertToDTO(reviewRepository.save(newReview));
+    }
+
+    private ReviewDto convertToDto(Review review) {
+        return reviewMapper.convertToDTO(review);
     }
 }
